Align issue 60 integration numbering and E2E coverage
