+++ e2e/README.md
@@ -78,9 +78,9 @@ created`, gegen eine gefüllte Datenbank bricht der Lauf also laut ab.
 | `src/controls.ts` | Sichtbarkeit und Überdeckung über `elementFromPoint` |
 | `src/layout.ts` | Seitenüberlauf und abgeschnittener Text — als Funktion der Fensterbreite |
 | `src/pseudoLocale.ts` | Verlängert im Browser jeden Text, der dem Cockpit gehört, um 35 % |
-| `tests/` | Die zehn Prüfdateien, in Ausführungsreihenfolge nummeriert |
+| `tests/` | Die zwölf Prüfdateien, in Ausführungsreihenfolge nummeriert |
 
-## Die zehn Prüfdateien
+## Die zwölf Prüfdateien
 
 | Datei | Prüfung |
 | --- | --- |
@@ -95,6 +95,7 @@ created`, gegen eine gefüllte Datenbank bricht der Lauf also laut ab.
 | `09-i18n-layout.spec.ts` | Layout in Deutsch, Englisch und Pseudo-Locale bei 1920, 1440 und 1280 |
 | `10-architecture-focus.spec.ts` | Architektur-Fokus, exakte Pane-Wiederherstellung und Overlay-Geometrie in Deutsch und Englisch |
 | `11-canvas-hit-areas.spec.ts` | Screen-space Hit-Areas bei mehreren Zoomstufen und Fine/Coarse Pointer |
+| `12-spatial-keyboard-navigation.spec.ts` | Vollständiger Arrow-only Keyboard-Walk, Pan ohne Zoom und Fokus-Fallback |
 
 `01`–`08` sind die verbindlichen Prüfungen des v0-Epics (#14) und bleiben
 unverändert: dieselben Assertions, dieselben 1920 × 1080, dieselbe Sprache.
